MakeFile refactor
