+++ etapa3/Makefile
@@ -1,5 +1,20 @@
-etapa3: y.tab.c lex.yy.c main.c
-	gcc -o etapa3 lex.yy.c y.tab.c main.c -ll
+etapa3: y.tab.o lex.yy.o main.o hash.o astree.o
+	gcc -o etapa3 y.tab.o lex.yy.o main.o hash.o astree.o -ll
+
+y.tab.o: y.tab.c
+	gcc -c y.tab.c
+
+lex.yy.o: lex.yy.c
+	gcc -c lex.yy.c
+
+hash.o: hash.c
+	gcc -c hash.c
+
+astree.o: astree.c
+	gcc -c astree.c
+
+main.o: main.c
+	gcc -c main.c
 
 y.tab.c: parser.y
 	yacc -d -v parser.y
