+++ etapa3/hash.c
@@ -1,25 +1,10 @@
 #define _XOPEN_SOURCE 500 /* Enable certain library functions (strdup) on linux.  See feature_test_macros(7) */
 
+#include "hash.h"
 #include <stdlib.h>
 #include <stdio.h>
 #include <limits.h>
 #include <string.h>
-
-struct entry_s {
-	char *key;
-	char *value;
-	int type;
-	struct entry_s *next;
-};
-
-typedef struct entry_s entry_t;
-
-struct hashtable_s {
-	int size;
-	struct entry_s **table;
-};
-
-typedef struct hashtable_s hashtable_t;
 
 /* Create a new hashtable. */
 hashtable_t *ht_create( int size ) {
